[#5]Refact: divide severApi and serverStatus files
