+++ src/store/actions/serverApi.ts
@@ -1,17 +1,4 @@
-import {
-  START_LOADING,
-  FINISH_LOADING,
-  GET_NOTE_REQUEST,
-} from './actionstypes';
-
-export const startLoading = (requestType: string) => ({
-  type: START_LOADING,
-  requestType,
-});
-export const finishLoding = (requestType: string) => ({
-  type: FINISH_LOADING,
-  requestType,
-});
+import { GET_NOTE_REQUEST } from './actionstypes';
 
 export const getNoteSagaAction = () => {
   return {
